[MIG] mail_preview_audio: Migration to 15.0
TT41546
