+++ mail_preview_audio/README.rst
@@ -14,13 +14,13 @@ Preview audio files
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsocial-lightgray.png?logo=github
-    :target: https://github.com/OCA/social/tree/14.0/mail_preview_audio
+    :target: https://github.com/OCA/social/tree/15.0/mail_preview_audio
     :alt: OCA/social
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/social-14-0/social-14-0-mail_preview_audio
+    :target: https://translation.odoo-community.org/projects/social-15-0/social-15-0-mail_preview_audio
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/205/14.0
+    :target: https://runbot.odoo-community.org/runbot/205/15.0
     :alt: Try me on Runbot
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
@@ -38,7 +38,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/social/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/social/issues/new?body=module:%20mail_preview_audio%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/social/issues/new?body=module:%20mail_preview_audio%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -55,6 +55,9 @@ Contributors
 
 * Enric Tobella <etobella@creublanca.es>
 
+* `Tecnativa <https://www.tecnativa.com>`_:
+  * David Vidal
+
 Maintainers
 ~~~~~~~~~~~
 
@@ -68,6 +71,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/social <https://github.com/OCA/social/tree/14.0/mail_preview_audio>`_ project on GitHub.
+This module is part of the `OCA/social <https://github.com/OCA/social/tree/15.0/mail_preview_audio>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
