+++ mail_preview_audio/readme/CONTRIBUTORS.rst
@@ -1,1 +1,4 @@
 * Enric Tobella <etobella@creublanca.es>
+
+* `Tecnativa <https://www.tecnativa.com>`_:
+  * David Vidal
